split testing code into separate header and associated source files
